requirements.txt
python=3.7
tqdm=4.31.1
numpy=1.16.4
python-Levenshtein=0.12.0